+++ webgear-server/src/WebGear/Trait/Header.hs
@@ -1,10 +1,9 @@
-{-|
-Copyright        : (c) Raghu Kaippully, 2020
-License          : MPL-2.0
-Maintainer       : rkaippully@gmail.com
-
-Traits related to HTTP headers.
--}
+-- |
+-- Copyright        : (c) Raghu Kaippully, 2020
+-- License          : MPL-2.0
+-- Maintainer       : rkaippully@gmail.com
+--
+-- Traits related to HTTP headers.
 module WebGear.Trait.Header
   ( Header
   , HeaderFail (..)