+++ webgear-server/src/WebGear/Trait/Path.hs
@@ -1,10 +1,9 @@
-{-|
-Copyright        : (c) Raghu Kaippully, 2020
-License          : MPL-2.0
-Maintainer       : rkaippully@gmail.com
-
-Traits related to the route path of a request.
--}
+-- |
+-- Copyright        : (c) Raghu Kaippully, 2020
+-- License          : MPL-2.0
+-- Maintainer       : rkaippully@gmail.com
+--
+-- Traits related to the route path of a request.
 module WebGear.Trait.Path
   ( Path
   , PathVar