+++ webgear-server/ChangeLog.md
@@ -1,3 +1,8 @@
-# Changelog for WebGear
+# Changelog for webgear-server
 
 ## Unreleased changes
+
+### Added
+
+- Support basic traits and middlewares
+- Automated tests